Move classes to specific package
