+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/prefs/DeveloperOptionsViewModelTest.kt
@@ -3,6 +3,8 @@ package com.woocommerce.android.ui.prefs
 import androidx.lifecycle.SavedStateHandle
 import com.woocommerce.android.R
 import com.woocommerce.android.model.UiString
+import com.woocommerce.android.ui.prefs.developer.DeveloperOptionsRepository
+import com.woocommerce.android.ui.prefs.developer.DeveloperOptionsViewModel
 import com.woocommerce.android.util.captureValues
 import com.woocommerce.android.viewmodel.BaseUnitTest
 import kotlinx.coroutines.ExperimentalCoroutinesApi
